Trimmed code, refactored
