+++ Tabloid/Controllers/TagController.cs
@@ -1,13 +1,11 @@
-﻿using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using System.Collections.Generic;
-using Tabloid.Repositories;
-using Microsoft.VisualBasic;
-using System.Security.Claims;
-using Tabloid.Models.ViewModels;
+﻿using System.Collections.Generic;
+
 using Microsoft.AspNetCore.Authorization;
-using System;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
+
 using Tabloid.Models;
+using Tabloid.Repositories;
 
 namespace Tabloid.Controllers
 {
@@ -23,62 +21,11 @@ namespace Tabloid.Controllers
         }
 
         // GET: TagController
-        [HttpGet("Index")]
+        [HttpGet]
         public ActionResult Index()
         {
             var tags = _tagRepository.GetAllTags();
-            return Ok(_tagRepository.GetAllTags());
-        }
-
-        // POST: api/Tag
-        [Authorize(Roles = "Admin")]
-        [HttpPost("Create")]
-        public ActionResult Create(Tag tag)
-        {
-            try
-            {
-                _tagRepository.AddTag(tag);
-
-                return Ok(new { message = "Success" });
-            }
-            catch
-            {
-                return Ok(tag);
-            }
-        }
-
-        // POST: api/Tag/Edit/5
-        [Authorize(Roles = "Admin")]
-        [HttpPost("Edit/{id}")]
-        public ActionResult Edit(int id, Tag tag)
-        {
-            try
-            {
-                _tagRepository.UpdateTag(tag);
-
-                return Ok();
-            }
-            catch
-            {
-                return Ok(tag);
-            }
-        }
-
-        // POST: api/Tag/Delete/5
-        [Authorize(Roles = "Admin")]
-        [HttpPost("Delete/{id}")]
-        public ActionResult Delete(int id)
-        {
-            try
-            {
-                _tagRepository.DeletePostTagsByTag(id);
-
-                return Ok();
-            }
-            catch
-            {
-                return Ok();
-            }
+            return Ok(tags);
         }
 
     }
