+++ Tabloid/Repositories/ITagRepository.cs
@@ -6,13 +6,5 @@ namespace Tabloid.Repositories
     public interface ITagRepository
     {
         List<Tag> GetAllTags();
-        Tag GetTagById(int id);
-        List<int> GetTagsByPostId(int id);
-        List<Tag> GetTagsOnPost(int id);
-        void DeletePostTagsByPost(int postId);
-        void DeletePostTagsByTag(int tagId);
-        void AddTag(Tag tag);
-        void DeleteTag(Tag tag);
-        void UpdateTag(Tag tag);
     }
 }
