+++ Tabloid/client/src/components/Tags/Tag.js
